Update [GASW] : Refactor most classes into Spring framework

+++ src/main/java/fr/insalyon/creatis/gasw/execution/FailOver.java
@@ -40,7 +40,6 @@ import fr.insalyon.creatis.gasw.GaswUtil;
 import fr.insalyon.creatis.gasw.bean.DataToReplicate;
 import fr.insalyon.creatis.gasw.bean.SEEntryPoint;
 import fr.insalyon.creatis.gasw.dao.DAOException;
-import fr.insalyon.creatis.gasw.dao.DAOFactory;
 import fr.insalyon.creatis.gasw.dao.DataToReplicateDAO;
 import java.io.BufferedReader;
 import java.io.Closeable;
@@ -52,62 +51,52 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
+import java.util.concurrent.TimeUnit;
+
+import fr.insalyon.creatis.gasw.dao.SEEntryPointsDAO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Service;
 
-public class FailOver extends Thread {
+@Service
+public class FailOver {
 
-    private static final Logger logger = LoggerFactory.getLogger(FailOver.class);
-    private static FailOver instance;
-    private volatile boolean stop = false;
-    private DataToReplicateDAO dataToReplicateDAO;
+    private final Logger logger = LoggerFactory.getLogger(getClass());
 
-    public synchronized static FailOver getInstance() {
-        if (instance == null) {
-            instance = new FailOver();
-            instance.start();
-        }
-        return instance;
+    private final GaswConfiguration config;
+    private final DataToReplicateDAO dataToReplicateDAO;
+    private final SEEntryPointsDAO seEntryPointDAO;
+
+    public FailOver(GaswConfiguration config, DataToReplicateDAO dataToReplicateDAO, SEEntryPointsDAO seEntryPointDAO) {
+        this.config = config;
+        this.dataToReplicateDAO = dataToReplicateDAO;
+        this.seEntryPointDAO = seEntryPointDAO;
     }
 
-    private FailOver() {
+    @Scheduled(fixedDelayString = "${gasw.default.sleep-time}", timeUnit = TimeUnit.SECONDS)
+    private void run() {
+        if (!config.isFailOverEnabled()) { return; }
         try {
-            dataToReplicateDAO = DAOFactory.getDAOFactory().getDataToReplicateDAO();
-        } catch (DAOException ex) {
-            logger.error("Unable to start Fail Over thread.");
-        }
-    }
+            for (DataToReplicate data : dataToReplicateDAO.get()) {
+                try {
+                    replicate(data.getUrl());
+                    dataToReplicateDAO.remove(data);
 
-    @Override
-    public void run() {
-        try {
-            while (!stop) {
+                } catch (GaswException ex) {
 
-                for (DataToReplicate data : dataToReplicateDAO.get()) {
-                    try {
-                        replicate(data.getUrl());
+                    if (data.getRetries() + 1 < config.getFailOverMaxRetry()) {
+                        data.setRetries(data.getRetries() + 1);
+                        data.setEventDate(new Date());
+                        dataToReplicateDAO.update(data);
+                    } else {
+                        logger.warn("Achieved data max attempts to reply '{}'.", data.getUrl().getPath());
                         dataToReplicateDAO.remove(data);
-
-                    } catch (GaswException ex) {
-
-                        if (data.getRetries() + 1 < GaswConfiguration.getInstance().getFailOverMaxRetry()) {
-                            data.setRetries(data.getRetries() + 1);
-                            data.setEventDate(new Date());
-                            dataToReplicateDAO.update(data);
-                        } else {
-                            logger.warn("Achieved data max attempts to reply '{}'.", data.getUrl().getPath());
-                            dataToReplicateDAO.remove(data);
-                        }
                     }
                 }
-                Thread.sleep(GaswConfiguration.getInstance().getDefaultSleeptime());
             }
         } catch (DAOException ex) {
-            // do nothing
-        } catch (GaswException ex) {
-            // do nothing
-        } catch (InterruptedException ex) {
-            logger.error("InterruptedException: ",ex);
+            logger.error("DAOException: ",ex);
         }
     }
 
@@ -130,16 +119,12 @@ public class FailOver extends Thread {
         }
     }
 
-    public synchronized void terminate() {
-        this.stop = true;
-    }
-
     private void replicate(URI uri) throws GaswException {
 
         List<URI> replicas = getReplicas(uri);
 
         for (URI replica : replicas) {
-            if (replica.getHost().equals(GaswConfiguration.getInstance().getFailOverHost())) {
+            if (replica.getHost().equals(config.getFailOverHost())) {
                 return;
             }
         }
@@ -239,19 +224,19 @@ public class FailOver extends Thread {
         return replicas;
     }
 
-    private String getDestinationSURL() throws GaswException {
+    private String getDestinationSURL() {
 
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
 
-        return "srm://" + GaswConfiguration.getInstance().getFailOverHost()
-                + ":" + GaswConfiguration.getInstance().getFailOverPort()
-                + "/srm/managerv2?SFN=" + GaswConfiguration.getInstance().getFailOverHome()
+        return "srm://" + config.getFailOverHost()
+                + ":" + config.getFailOverPort()
+                + "/srm/managerv2?SFN=" + config.getFailOverHome()
                 + "/" + sdf.format(new Date()) + "/file-" + UUID.randomUUID();
     }
 
     private String[] getSourceTypeAndSURL(String host, String path) throws DAOException {
 
-        SEEntryPoint ep = DAOFactory.getDAOFactory().getSEEntryPointDAO().getByHostName(host);
+        SEEntryPoint ep = seEntryPointDAO.getByHostName(host);
         String[] source = new String[]{
             ep.getHome().contains("managerv1") ? "srmv1" : "srmv2",
             "srm://" + ep.getId().getHostname() + ":" + ep.getId().getPort()
